+++ changelog.md
@@ -337,6 +337,8 @@
 
 - Added `dom.scrollIntoView` proc with options
 
+- Added `ControlChars`, `GraphicChars`, `PrintableChars`, and `Punctuation` sets to `strutils`.
+
 ## Language changes
 
 - `nimscript` now handles `except Exception as e`.